+++ CHANGELOG.rst
@@ -18,6 +18,18 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
       (only in comment or documentation).
 
 
+v1.2.4 (unreleased)
+===================
+
+Bug fix release
+
+Other
+-----
+
+- Add missing docstring to the classes :class:`~deprecated.classic.ClassicAdapter`
+  and :class:`~deprecated.sphinx.SphinxAdapter`.
+
+
 v1.2.3 (2018-09-12)
 ===================